jianpu4.ly from Paul Morris.
Ok, the problem was that your music file was for LilyPond 2.19 and jianpu3.ly only works with 2.18.

Attached is jianpu4.ly which works with either 2.18 or 2.19.  It also creates a custom Staff context called “StaffJianpu" so you can just write:

\new StaffJianpu {
  % music etc. goes here just like
  % with the standard Staff context
}

jianpu4.ly is ready to be included in files.  (Feel free to rename it if you like.)  I moved the test music to jianpu-test.ly.

http://lilypond.1069038.n5.nabble.com/Jianpu-Notation-tp174685p175085.html