jianpu5.ly from Paul Morris.
I see now that the octave dots in Jianpu are not based on the absolute octave of the note but are relative to the current key / key signature.  In other words, the Jianpu octave increases between 7 and 1, and not between b and c.

Attached is a new version jianpu5.ly that gets the octave dots right.  Let me know if you see anything else that’s not correct.

I have also changed the name of the custom staff context to JianpuStaff to match the other staff contexts in LilyPond.  (Although I have a slight preference for having it the other way around, consistency is good.)

http://lilypond.1069038.n5.nabble.com/Jianpu-Notation-tp174685p175206.html