jianpu6b.ly from Paul Morris.
Hello,

> On Apr 24, 2015, at 12:43 AM, Paul Morris <[hidden email]> wrote:
>
>> > I just spotted \time 3/4 or 4/4, jianpu display " 6 - ." , I think it should be "6 - -".
>> Indeed, in Jianpu, dots follow only by quarter notes and below.
>
> Yes, this is a “known issue”.  I have started working on it, but it’s a bit beyond the things I am familiar with and already know about.  (I have gotten it to add the right number of dashes, but the dots are still being displayed when they shouldn’t.  Work for another day…)

I’m attaching a new version that fixes this.

However, I see now that really a different approach is needed.  Basically whole notes and half notes should be converted into the equivalent number of quarter notes, with the first one displayed as a number and the rest as dashes.  That way the spacing will be much better.

I’ve started experimenting with this.  It will take using a music function.  This will also allow automatically inserting markups for the octave dots, which will then make it easier to have beams that span more than one note.  So there are several benefits to this approach.

http://lilypond.1069038.n5.nabble.com/Jianpu-Notation-tp174685p175363.html