jianpu9.ly from Paul Morris.
Thanks Ming for catching that glitch.  Everything worked fine for simple test cases with rests, but somehow for those rests in your score the code was somehow registering a pitch for those rest events (I don’t know why…) so they were being treated as notes.

The new version 9 that's attached solves this problem.  Not hard to fix once I figured out the cause (just check to see if should be a rest first, rather than first checking for a pitch as an indication that it should be a note).  Let me know if you find any other odd glitches like this.

Oh, and thanks Werner, I agree that it has gotten a lot better!

http://lilypond.1069038.n5.nabble.com/Jianpu-Notation-tp174685p175682.html